fix: test direstory
