+++ scrapper/src/test/java/edu/java/IntegrationTest.java
@@ -1,4 +1,4 @@
-package edu.java.integration;
+package edu.java;
 
 import java.io.File;
 import java.sql.DriverManager;
