+++ scrapper/src/test/java/edu/java/JooqLinkUpdaterTest.java
@@ -1,9 +1,8 @@
-package edu.java.integration;
+package edu.java;
 
 import com.github.tomakehurst.wiremock.junit5.WireMockExtension;
-import com.github.tomakehurst.wiremock.junit5.WireMockTest;
-import edu.java.service.jpa.JpaLinkService;
-import edu.java.service.jpa.JpaLinkUpdater;
+import edu.java.service.jooq.JooqLinkService;
+import edu.java.service.jooq.JooqLinkUpdater;
 import edu.java.service.model.EventLink;
 import edu.java.service.model.EventName;
 import java.io.IOException;
@@ -17,6 +16,7 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.jooq.JooqAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.Rollback;
 import org.springframework.test.context.jdbc.Sql;
@@ -28,19 +28,24 @@ import static com.github.tomakehurst.wiremock.client.WireMock.urlPathMatching;
 import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMockConfig;
 import static org.assertj.core.api.Assertions.assertThat;
 
-@SpringBootTest(classes = {JpaTestConfig.class})
+@SpringBootTest(classes = {JooqTestConfig.class, JooqAutoConfiguration.class})
 @Sql(value = "classpath:sql/put_chat.sql", executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
-@WireMockTest
-public class JpaUpdaterServiceTest extends IntegrationTest {
-    @Autowired JpaLinkService linkService;
-    @Autowired JpaLinkUpdater linkUpdater;
-
+public class JooqLinkUpdaterTest extends IntegrationTest {
+    private final JooqLinkService linkService;
+    private final JooqLinkUpdater linkUpdater;
     private final int offset = 365;
+
+    @Autowired
+    public JooqLinkUpdaterTest(JooqLinkService linkService, JooqLinkUpdater linkUpdater) {
+        this.linkService = linkService;
+        this.linkUpdater = linkUpdater;
+    }
 
     @RegisterExtension
     static WireMockExtension wm = WireMockExtension.newInstance()
         .options(wireMockConfig().port(8080))
         .build();
+
     @Test
     @Transactional
     @Rollback
@@ -51,7 +56,7 @@ public class JpaUpdaterServiceTest extends IntegrationTest {
         wm.stubFor(get(urlPathMatching("/repos/Marijarin/tocook/events"))
             .willReturn(aResponse()
                 .withHeader("Content-Type", "application/json")
-                .withBody(Files.readString(Path.of("src/test/resources/json/gh-issue.json")))
+                .withBody(Files.readString(Path.of("src/test/resources/json/gh-create.json")))
                 .withStatus(200)));
         try {
             linkDao.setLastUpdated(OffsetDateTime.now().minus(Duration.ofDays(offset)));
@@ -61,7 +66,7 @@ public class JpaUpdaterServiceTest extends IntegrationTest {
             System.out.println(e.getMessage());
         }
         assertThat(eventList.getFirst().getEvent().getDescription())
-            .isEqualTo(EventName.getEventMap().get("IssueCommentEvent")
+            .isEqualTo(EventName.getEventMap().get("CreateEvent")
                 .getDescription());
         wm.verify(getRequestedFor(urlPathMatching("/repos/.*")));
     }
@@ -76,7 +81,7 @@ public class JpaUpdaterServiceTest extends IntegrationTest {
         wm.stubFor(get(urlPathMatching("/questions/1/timeline.*"))
             .willReturn(aResponse()
                 .withHeader("Content-Type", "application/json")
-                .withBody(Files.readString(Path.of("src/test/resources/json/sof-comment.json")))
+                .withBody(Files.readString(Path.of("src/test/resources/json/sof-answer.json")))
                 .withStatus(200)));
         try {
             linkDao.setLastUpdated(OffsetDateTime.now().minus(Duration.ofDays(offset)));
@@ -86,7 +91,7 @@ public class JpaUpdaterServiceTest extends IntegrationTest {
             System.out.println(e.getMessage());
         }
         assertThat(eventList.getFirst().getEvent().getDescription())
-            .isEqualTo(EventName.getEventMap().get("comment")
+            .isEqualTo(EventName.getEventMap().get("answer")
                 .getDescription());
         wm.verify(getRequestedFor(urlPathMatching("/questions/1/timeline.*")));
     }
