+++ scrapper/src/test/java/edu/java/JpaLinkServiceTest.java
@@ -1,23 +1,22 @@
-package edu.java.integration;
+package edu.java;
 
-import edu.java.service.jooq.JooqLinkService;
+import edu.java.service.jpa.JpaLinkService;
 import java.net.URI;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.jooq.JooqAutoConfiguration;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.Rollback;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.transaction.annotation.Transactional;
 import static org.assertj.core.api.Assertions.assertThat;
 
-@SpringBootTest(classes = {JooqTestConfig.class, JooqAutoConfiguration.class})
+@SpringBootTest(classes = {JpaTestConfig.class})
 @Sql(value = "classpath:sql/put_chat.sql", executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
-public class JooqLinkServiceTest extends IntegrationTest {
-    private final JooqLinkService linkService;
+public class JpaLinkServiceTest extends IntegrationTest {
+    private final JpaLinkService linkService;
 
     @Autowired
-    public JooqLinkServiceTest(JooqLinkService linkService) {
+    public JpaLinkServiceTest(JpaLinkService linkService) {
         this.linkService = linkService;
     }
 
@@ -35,11 +34,10 @@ public class JooqLinkServiceTest extends IntegrationTest {
         assertThat(result.getUri()).isEqualTo(uri);
         assertThat(listResult.stream().toList().getFirst().getUri()).isEqualTo(uri);
     }
-
     @Test
     @Transactional
     @Rollback
-    void removesAndNotFindsLink() {
+    void removesAndNotFindsLink(){
         var chatId = 10000L;
         var uri = URI.create(
             "https://stackoverflow.com/questions/68547843/bean-named-userservice-is-expected-to-be-of-type-demo-spring-orm-userservice");
@@ -50,5 +48,4 @@ public class JooqLinkServiceTest extends IntegrationTest {
 
         assertThat(listResult.isEmpty()).isTrue();
     }
-
 }
