+++ scrapper/src/test/java/edu/java/TableMigrationsIntegrationTest.java
@@ -1,4 +1,4 @@
-package edu.java.integration;
+package edu.java;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
